Split read_scores into read_scores and extract_scores. Update tests.
read_scores just takes an open file (or other buffer) and reads scores from it,
while extract_scores takes a path and opens the scorefile, passing that open
file to read_scores to do the work.

It's possible that I'll remove extract_scores from the library later, but I'm
leaving it for now, while I'm still in the exploratory programming stage.